# backend/API/openapi.yml
openapi: 3.1.0
info:
  title: Training Management API
  description: API for managing training sessions, students, trainers, schedules, and plans
  version: v1.0.0
servers:
- url: http://localhost:8080
  description: Local development server
security:
- bearerAuth: []
paths:
...
  /administrators:
    get:
      tags:
      - people
      operationId: findAllAdministrators
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListAdministratorsDTO'
    post:
      tags:
      - people
      operationId: createAdministrator
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdministratorDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ListAdministratorsDTO'
  /administrators/{administratorId}:
    get:
      tags:
      - people
      operationId: findAdministratorById
      parameters:
      - name: administratorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ListAdministratorsDTO'
    put:
      tags:
      - people
      operationId: updateAdministrator
      parameters:
      - name: administratorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdministratorDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdministratorResponseDTO'
    delete:
      tags:
      - people
      operationId: deleteAdministrator
      parameters:
      - name: administratorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
    patch:
      tags:
      - people
      operationId: patchAdministrator
      parameters:
      - name: administratorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchAdministratorRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdministratorResponseDTO'
  /administrators/{administratorId}/restore:
    patch:
      tags:
      - people
      operationId: restoreAdministrator
      parameters:
      - name: administratorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdministratorResponseDTO'
  /administrators/user-profile/{userId}:
    get:
      tags:
      - people
      operationId: findAdministratorByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ListAdministratorsDTO'
  /administrators/paginated:
    get:
      tags:
      - people
      operationId: findAllPaginated
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: includeInactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageListAdministratorsDTO'
...
  /events:
    get:
      tags:
      - events
      operationId: findAllEvents
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: includeInactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDTO'
    post:
      tags:
      - events
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
  /events/{eventId}:
    get:
      tags:
      - events
      operationId: findEventById
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
    put:
      tags:
      - events
      operationId: updateEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
    delete:
      tags:
      - events
      operationId: deleteEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
    patch:
      tags:
      - events
      operationId: patchEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchEventRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
...
  /events/{eventId}/participants/{studentId}:
    post:
      tags:
      - events
      operationId: addParticipant
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
    delete:
      tags:
      - events
      operationId: removeParticipant
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /events/{eventId}/restore:
    patch:
      tags:
      - events
      operationId: restoreEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
  /events/{eventId}/participants/{studentId}/attendance:
    patch:
      tags:
      - events
      operationId: toggleAttendance
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
...
components:
  schemas:
...
    EventRequestDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 0
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
        trainerId:
          type: string
          format: uuid
        participantIds:
          type: array
          items:
            type: string
            format: uuid
      required:
      - date
      - name
      - trainerId
    EventParticipantResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatar:
          type: string
        enrolledAt:
          type: string
          format: date-time
        present:
          type: boolean
    EventResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        description:
          type: string
        instructorId:
          type: string
          format: uuid
        instructor:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/EventParticipantResponseDTO'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
    CreateAdministratorDTO:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 100
          minLength: 0
        lastName:
          type: string
          maxLength: 100
          minLength: 0
        email:
          type: string
          maxLength: 255
          minLength: 0
        password:
          type: string
          maxLength: 255
          minLength: 6
      required:
      - email
      - firstName
      - lastName
      - password
    AdministratorResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        fullName:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ListAdministratorsDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        fullName:
          type: string
        active:
          type: boolean
        joinDate:
          type: string
          format: date-time
...
    PatchEventRequestDTO:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
        trainerId:
          type: string
          format: uuid
        participantIds:
          type: array
          items:
            type: string
            format: uuid
    PatchAdministratorRequestDTO:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 100
          minLength: 0
        lastName:
          type: string
          maxLength: 100
          minLength: 0
        email:
          type: string
          maxLength: 255
          minLength: 0
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageListAdministratorsDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ListAdministratorsDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT